Fix: Catalog creation / deletion for motherduck

+++ sqlmesh/core/engine_adapter/duckdb.py
@@ -47,16 +47,30 @@ class DuckDBEngineAdapter(LogicalMergeMixin, GetCurrentCatalogFromFunctionMixin,
         self.execute(exp.Use(this=exp.to_identifier(catalog)))
 
     def _create_catalog(self, catalog_name: exp.Identifier) -> None:
-        db_filename = f"{catalog_name.output_name}.db"
-        self.execute(
-            exp.Attach(this=exp.alias_(exp.Literal.string(db_filename), catalog_name), exists=True)
-        )
+        if not self._is_motherduck:
+            db_filename = f"{catalog_name.output_name}.db"
+            self.execute(
+                exp.Attach(
+                    this=exp.alias_(exp.Literal.string(db_filename), catalog_name), exists=True
+                )
+            )
+        else:
+            self.execute(
+                exp.Create(this=exp.Table(this=catalog_name), kind="DATABASE", exists=True)
+            )
 
     def _drop_catalog(self, catalog_name: exp.Identifier) -> None:
-        db_file_path = Path(f"{catalog_name.output_name}.db")
-        self.execute(exp.Detach(this=catalog_name, exists=True))
-        if db_file_path.exists():
-            db_file_path.unlink()
+        if not self._is_motherduck:
+            db_file_path = Path(f"{catalog_name.output_name}.db")
+            self.execute(exp.Detach(this=catalog_name, exists=True))
+            if db_file_path.exists():
+                db_file_path.unlink()
+        else:
+            self.execute(
+                exp.Drop(
+                    this=exp.Table(this=catalog_name), kind="DATABASE", cascade=True, exists=True
+                )
+            )
 
     def _df_to_source_queries(
         self,
@@ -198,3 +212,7 @@ class DuckDBEngineAdapter(LogicalMergeMixin, GetCurrentCatalogFromFunctionMixin,
                 expr.sql(dialect=self.dialect) for expr in partitioned_by_exps
             )
             self.execute(f"ALTER TABLE {table_name_str} SET PARTITIONED BY ({partitioned_by_str});")
+
+    @property
+    def _is_motherduck(self) -> bool:
+        return self._extra_config.get("is_motherduck", False)
